FileServe: lets see if it works again Youtube: fixed unescape, fixes #4191
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@15735 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/hoster/Youtube.java
@@ -81,7 +81,7 @@ public class Youtube extends PluginForHost {
                         }
                     }
                     i--;
-                    sb.append((char) Short.parseShort(sb2.toString(), 16));
+                    sb.append((char) Long.parseLong(sb2.toString(), 16));
                     continue;
                 case 'x':
                     /* normal hex coding */
@@ -93,7 +93,7 @@ public class Youtube extends PluginForHost {
                         sb2.append(ch);
                     }
                     i--;
-                    sb.append((char) Short.parseShort(sb2.toString(), 16));
+                    sb.append((char) Long.parseLong(sb2.toString(), 16));
                     continue;
                 default:
                     sb.append(ch);
